+++ frontends/landlord/src/app/parkingSpots/list.tsx
@@ -4,9 +4,9 @@ import React, { useState } from 'react'
 import { FlatList, View } from 'react-native'
 import { ActivityIndicator, Avatar, Button, ButtonProps, Card, Text, useTheme } from 'react-native-paper'
 import { CoreClientBuilder } from 'src/apiClient/CoreClientBuilder'
-import { useCoreClient } from 'src/apiClient/useCoreClient'
 import { Screen } from 'src/components/Screen'
 import { useAuthContext } from 'src/contexts/AuthContext'
+import { useCoreClient } from 'src/hooks/useCoreClient'
 import { useCounter } from 'src/hooks/useCounter'
 import { useNavigationHeader } from 'src/hooks/useNavigationHeader'
 import { showErrorToast } from 'src/toasts/showErrorToast'
